Move fetching (enrolled) clubs to actions and background jobs
